refactor(NotificationBar): client専用処理をclient/MessageAreaへ切り出しNotificationBar自体から'use client'を除去する
ariaNotify呼び出しやMutationObserverなど、client専用APIに依存する
callbackRefのロジックはDOM要素にアタッチされるMessageArea側の関心事
であり、NotificationBar本体には不要だった。

このロジックをclient/MessageArea.tsxへ移動することで、NotificationBar
自体はuseMemoのみを使うモジュールになり'use client'を除去できる。
実際にビルド出力でNotificationBar.jsに'use client'が付かなくなり、
client/MessageArea.jsにのみ付くことを確認済み。

hookが存在せずindex.tsからも公開されない内部コンポーネントのため、
client/直下にフラット配置する（CLAUDE.mdの判断基準）。
