+++ packages/smarthr-ui/src/components/NotificationBar/NotificationBar.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import {
   type ComponentProps,
   type ComponentPropsWithoutRef,
@@ -7,28 +5,17 @@ import {
   Fragment,
   type PropsWithChildren,
   type ReactNode,
-  memo,
-  useCallback,
   useMemo,
 } from 'react'
 import { type VariantProps, tv } from 'tailwind-variants'
 
-import { useCallbackRefCleanupForReact18 } from '../../hooks/client/useCallbackRefCleanupForReact18'
-import { useLatest } from '../../hooks/useLatest'
 import { Localizer } from '../../intl'
 import { Button } from '../Button'
-import {
-  FaCircleCheckIcon,
-  FaCircleExclamationIcon,
-  FaCircleInfoIcon,
-  FaRotateIcon,
-  FaTriangleExclamationIcon,
-  FaXmarkIcon,
-  WarningIcon,
-} from '../Icon'
+import { FaXmarkIcon } from '../Icon'
 import { Cluster } from '../Layout'
 import { Panel } from '../Panel'
-import { Text } from '../Text'
+
+import { MessageArea } from './client'
 
 const classNameGenerator = tv({
   slots: {
@@ -157,25 +144,6 @@ type Props = PanelLayerProps &
   Omit<ComponentPropsWithoutRef<'div'>, keyof PanelLayerProps> &
   Omit<BaseProps, keyof PanelLayerProps>
 
-const ABSTRACT_ICON_MAPPER = {
-  info: FaCircleInfoIcon,
-  success: FaCircleCheckIcon,
-  error: FaCircleExclamationIcon,
-  sync: FaRotateIcon,
-}
-const ICON_MAPPER = {
-  normal: {
-    ...ABSTRACT_ICON_MAPPER,
-    warning: WarningIcon,
-  },
-  bold: {
-    ...ABSTRACT_ICON_MAPPER,
-    warning: FaTriangleExclamationIcon,
-  },
-} as const
-
-const ROLE_STATUS_TYPE_REGEX = /^(info|sync|success)$/
-
 export const NotificationBar: FC<Props> = ({
   type,
   bold,
@@ -216,52 +184,11 @@ export const NotificationBar: FC<Props> = ({
     }
   }, [animate, base, bold, type, className])
 
-  const latest = useLatest({ role, type })
-
-  const callbackRef = useCallbackRefCleanupForReact18(
-    useCallback(
-      (node: HTMLElement | null) => {
-        if (!node) {
-          return
-        }
-
-        const ariaNotifyAction = () => {
-          const message = node.innerText || ''
-
-          if (!message) {
-            return
-          }
-
-          const actualRole =
-            latest.role || (ROLE_STATUS_TYPE_REGEX.test(latest.type) ? 'status' : 'alert')
-
-          document.ariaNotify(message, {
-            priority: actualRole === 'alert' ? 'high' : 'normal',
-          })
-        }
-
-        ariaNotifyAction()
-
-        const observer = new MutationObserver(ariaNotifyAction)
-        observer.observe(node, {
-          childList: true,
-          subtree: true,
-          characterData: true,
-        })
-
-        return () => {
-          observer.disconnect()
-        }
-      },
-      [latest],
-    ),
-  )
-
   return (
     <WrapBase {...baseProps}>
       <div {...rest} className={classNames.wrapper}>
         <Cluster gap={1} align="center" justify="flex-end" className={classNames.inner}>
-          <MessageArea callbackRef={callbackRef} type={type} bold={bold} classNames={classNames}>
+          <MessageArea role={role} type={type} bold={bold} classNames={classNames}>
             {children}
           </MessageArea>
           {subActionArea && (
@@ -287,25 +214,4 @@ export const NotificationBar: FC<Props> = ({
   )
 }
 
-const MessageArea = memo<
-  Pick<Props, 'children' | 'bold' | 'type'> & {
-    callbackRef: (node: HTMLElement | null) => void
-    classNames: { messageArea: string; icon: string }
-  }
->(({ callbackRef, children, bold, type, classNames }) => {
-  const Icon = ICON_MAPPER[bold ? 'bold' : 'normal'][type]
-
-  return (
-    <Text
-      as="div"
-      ref={callbackRef}
-      className={classNames.messageArea}
-      icon={{
-        prefix: <Icon className={classNames.icon} />,
-        gap: 0.5,
-      }}
-    >
-      {children}
-    </Text>
-  )
-})
+export type NotificationBarProps = Props
