+++ packages/smarthr-ui/src/components/NotificationBar/client/index.ts
@@ -1,0 +1,1 @@
+export { MessageArea } from './MessageArea'
